% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doas.R
\name{bayesian_conjugate_estimate}
\alias{bayesian_conjugate_estimate}
\title{Calculate Bayesian Conjugate Double Observer Estimate}
\usage{
bayesian_conjugate_estimate(both, obs1_only, obs2_only, iter = 10000)
}
\arguments{
\item{both}{Integer vector of numbers of groups seen by both observers}

\item{obs1_only}{Integer vector of numbers of groups seen by Observer 1 only}

\item{obs2_only}{Integer vector of numbers of groups seen by Observer 2 only}

\item{iter}{Integer number of Monte Carlo iterations (default 10000)}
}
\value{
A list containing the posterior 'mean' estimate vector and 'var' variance vector.
}
\description{
Uses Monte Carlo simulation from Beta posterior distributions (assuming Beta(1,1) priors)
to estimate the mean group abundance and its variance.
}
