% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doas.R
\name{chapman_estimate}
\alias{chapman_estimate}
\title{Calculate Chapman-adjusted Lincoln-Petersen estimate}
\usage{
chapman_estimate(both, obs1_only, obs2_only)
}
\arguments{
\item{both}{Integer vector of numbers of groups seen by both observers}

\item{obs1_only}{Integer vector of numbers of groups seen by Observer 1 only}

\item{obs2_only}{Integer vector of numbers of groups seen by Observer 2 only}
}
\value{
A numeric vector of the estimated total number of groups ($G_{est}$)
}
\description{
Calculate Chapman-adjusted Lincoln-Petersen estimate
}
