% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doas.R
\name{chapman_variance}
\alias{chapman_variance}
\title{Calculate Variance of the Chapman estimator}
\usage{
chapman_variance(both, obs1_only, obs2_only)
}
\arguments{
\item{both}{Integer vector of numbers of groups seen by both observers}

\item{obs1_only}{Integer vector of numbers of groups seen by Observer 1 only}

\item{obs2_only}{Integer vector of numbers of groups seen by Observer 2 only}
}
\value{
A numeric vector describing the variance of the estimated number of groups
}
\description{
Calculate Variance of the Chapman estimator
}
