% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doas.R
\name{dependent_zippin_estimate}
\alias{dependent_zippin_estimate}
\title{Calculate Dependent Double Observer (Zippin) Estimate}
\usage{
dependent_zippin_estimate(both, obs1_only, obs2_only)
}
\arguments{
\item{both}{Integer vector of numbers of groups seen by both observers}

\item{obs1_only}{Integer vector of numbers of groups seen by Observer 1 (Primary) only}

\item{obs2_only}{Integer vector of numbers of groups seen by Observer 2 (Secondary) only}
}
\value{
A numeric vector of the estimated total number of groups ($G_{est}$)
}
\description{
Calculate Dependent Double Observer (Zippin) Estimate
}
