% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doas.R
\name{dependent_zippin_variance}
\alias{dependent_zippin_variance}
\title{Calculate Variance of the Dependent Double Observer (Zippin) estimator}
\usage{
dependent_zippin_variance(both, obs1_only, obs2_only)
}
\arguments{
\item{both}{Integer vector of numbers of groups seen by both observers}

\item{obs1_only}{Integer vector of numbers of groups seen by Observer 1 (Primary) only}

\item{obs2_only}{Integer vector of numbers of groups seen by Observer 2 (Secondary) only}
}
\value{
A numeric vector describing the variance of the estimated number of groups
}
\description{
Calculate Variance of the Dependent Double Observer (Zippin) estimator
}
