% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaic.R
\name{QAIC.doas}
\alias{QAIC.doas}
\title{QAIC method for DOAS models}
\usage{
\method{QAIC}{doas}(object, ...)
}
\arguments{
\item{object}{A fitted `doas` object.}

\item{...}{Optionally more fitted `doas` objects.}
}
\value{
A numeric QAIC value, or a data frame if multiple models are passed.
}
\description{
Extracts the QAIC for DOAS models that used a covariate formula.
Standard AIC is undefined for the quasi-binomial GLM used under the hood,
so QAIC penalizes the binomial likelihood by the estimated overdispersion
parameter (c-hat).
}
\examples{
\dontrun{
mod1 <- doas(data, formula = ~ effort)
mod2 <- doas(data, formula = ~ effort + habitat)
QAIC(mod1, mod2)
}
}
