% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_pop}
\alias{simulate_pop}
\title{Simulate a Population}
\usage{
simulate_pop(pop_size, mean_group, tot_areas, area_sizes = NULL)
}
\arguments{
\item{pop_size}{Average total population size (Poisson distributed).}

\item{mean_group}{Average group size (Poisson distributed).}

\item{tot_areas}{Total number of survey sectors.}

\item{area_sizes}{Vector of probabilities/sizes for assigning groups to sectors. If NULL, uniform probabilities are used.}
}
\value{
A list containing the total population size and a list of groups per sector.
}
\description{
Simulates a population of individuals split into groups across multiple sectors.
}
