% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_survey}
\alias{simulate_survey}
\title{Simulate a Survey and format for DOAS}
\usage{
simulate_survey(
  sim_pop,
  det_p1,
  intra_group_det,
  det_group,
  obs_effect,
  detvar,
  surv_areas
)
}
\arguments{
\item{sim_pop}{A simulated population from \code{\link{simulate_pop}}.}

\item{det_p1}{Base detection probability for Observer 1.}

\item{intra_group_det}{Mean factor for count errors (e.g., 0.9 for undercounting).}

\item{det_group}{Base detection factor for group size scaling.}

\item{obs_effect}{Observer 2 relative efficiency (e.g., 0.75 means Obs 2 is 75% as effective).}

\item{detvar}{Standard deviation for detection probability across sectors.}

\item{surv_areas}{Number of sectors to receive Double Observer surveys.}
}
\value{
A list containing the total population size and a list of groups per sector.
}
\description{
Simulate a Survey and format for DOAS
}
