% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doas.R
\name{summary.doas}
\alias{summary.doas}
\title{Summary for doas objects}
\usage{
\method{summary}{doas}(object, ...)
}
\arguments{
\item{object}{An object of class `doas`}

\item{...}{Additional arguments}
}
\description{
Summary for doas objects
}
